requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
flake8==5.0.4
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
pycodestyle==2.9.1
pyflakes==2.5.0
python-telegram-bot==13.15
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tomli==2.0.1
tornado==6.1
types-requests==2.28.11.5
types-urllib3==1.26.25.4
types-waitress==2.1.4.3
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
waitress==2.1.2
Werkzeug==2.2.2
